:truck: Move components w/ test + styling  files to individual folders
