+++ src/components/index.js
@@ -1,15 +1,15 @@
 import "bootstrap/dist/css/bootstrap.min.css";
 export { App } from "./app/App.js";
-export { ProjectCardContainer } from "./projects/ProjectCardContainer";
-export { ProjectCard } from "./projects/ProjectCard";
-export { ProjectCardItem } from "./projects/ProjectCardItem";
+export { ProjectCardContainer } from "./projectCardContainer/ProjectCardContainer";
+export { ProjectCard } from "./projectCard/ProjectCard";
+export { ProjectCardItem } from "./projectCardItem/ProjectCardItem";
 export { ProjectProcessContainer } from "./projectProcess/ProjectProcessContainer";
 export { ProjectProcessRoutes } from "./projectProcess/ProjectProcessRoutes";
 export { ProjectProcessLinks } from "./projectProcess/ProjectProcessLinks";
-export { ContactLinkList } from "./navBar/ContactLinkList";
+export { ContactLinkList } from "./contactLinkList/ContactLinkList";
 // export { AboutLink } from "./navBar/AboutLink";
 export { About } from "./about/About";
 export { Routes } from "./routes/Routes";
 export { NavBar } from "./navBar/NavBar";
-export { NavPageLinks } from "./navBar/NavPageLinks";
-export { ContactLinkItem } from "./navBar/ContactLinkItem";
+export { NavPageLinks } from "./navBarLinks/NavPageLinks";
+export { ContactLinkItem } from "./contactLinkItem/ContactLinkItem";
